pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-loop"
version = "0.1.0"
description = "AI agent that follows instructions"
readme = "README.md"
requires-python = ">=3.7"
license = {text = "MIT"}
dependencies = [
    "requests",
    "beautifulsoup4",
    "google-generativeai",
    "llama-cpp-python",
    "huggingface_hub",
]

[project.scripts]
agent = "agent:main"